Create list-all-items-in-category
Operators that can be used with multiple-row subqueries include IN, ALL, ANY, and EXISTS. The multi row operators IN, ANY, ALL must be used with single row operators as shown in the option B

Generally, a subquery in SQL is used to retrieve data that will be used in the outer query. In this scenario, the inner query is executed first. Then, its results are used to evaluate the outer query